+++ mongo/bounties/init.js
@@ -1,0 +1,1 @@
+db.bounties.createIndex({ title: 'text' });