+++ packages/react-app/src/pages/api/bounties/index.ts
@@ -1,22 +1,21 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import dbConnect from '../../../utils/dbConnect';
-import Bounty from '../../../models/Bounty';
+import { getFilters, getSort, getBounties } from '../../../services/bounty.service';
 
 export default async function handler(
 	req: NextApiRequest,
 	res: NextApiResponse
 ): Promise<void> {
 	const { method } = req;
+	const filters = getFilters(req.query);
+	const sort = getSort(req.query);
 
 	await dbConnect();
 
 	switch (method) {
 	case 'GET':
 		try {
-			const bounties = await Bounty.find({
-				status: ['Open', 'In-Progress', 'In-Review'],
-			});
-			/* find all bounties that aren't in draft or deleted */
+			const bounties = await getBounties(filters, sort);
 			res.status(200).json({ success: true, data: bounties });
 		} catch (error) {
 			res.status(400).json({ success: false });
@@ -27,3 +26,4 @@ export default async function handler(
 		break;
 	}
 }
+