+++ packages/react-app/src/pages/api/customers/index.ts
@@ -1,0 +1,25 @@
+import { NextApiRequest, NextApiResponse } from 'next';
+import dbConnect from '../../../utils/dbConnect';
+import * as service from '../../../services/customer.service';
+
+export default async function handler(
+	req: NextApiRequest,
+	res: NextApiResponse
+): Promise<void> {
+	await dbConnect();
+
+	switch (req.method) {
+	case 'GET':
+		try {
+			const data = await service.getCustomers();
+			res.status(200).json({ success: true, data });
+		} catch (error) {
+			res.status(400).json({ success: false });
+		}
+		break;
+	default:
+		res.status(400).json({ success: false });
+		break;
+	}
+}
+